Split up models into querier, user, and application.
