+++ src/server/server.go
@@ -43,7 +43,7 @@ func (s *EaseServer) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 func createRoutingMux(client *db.Client) *mux.Router {
 
 	// Set up the queriers and controllers.
-	userQuerier := models.NewUserQuerier(client.Session)
+	userQuerier := models.NewModelQuerier(client.Session)
 
 	usercontroller.Init(userQuerier)
 	helpers.Init(userQuerier)
